requirements.txt
guidance==0.0.64
openai==0.27.10
python-dotenv==1.0.0
streamlit==1.26.0
